requirement.txt
asgiref==3.2.10
autopep8==1.5.7
bcrypt==3.2.0
boto3==1.18.54
botocore==1.21.54
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.9
Django==3.1.1
django-amazon-ses==4.0.0
django-clear-cache==0.3
django-cors-headers==3.8.0
django-filter==2.4.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djongo==1.3.6
elasticsearch==7.15.0
elasticsearch-dsl==7.4.0
gunicorn==20.0.4
idna==3.3
jmespath==0.10.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
python-dateutil==2.8.2
python-decouple==3.4
python-dotenv==0.18.0
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.2.4
toml==0.10.2
urllib3==1.26.7